Add app test for starting energy request
