+++ source/B2BApi.AppTests/Fixtures/B2BApiAppFixture.cs
@@ -278,11 +278,13 @@ public class B2BApiAppFixture : IAsyncLifetime
     }
 
     public void EnsureAppHostUsesFeatureFlagValue(
-        bool useRequestWholesaleServicesOrchestration = false)
+        bool useRequestWholesaleServicesOrchestration = false,
+        bool useRequestAggregatedMeasureDataOrchestration = false)
     {
         AppHostManager.RestartHostIfChanges(new Dictionary<string, string>
         {
             { $"FeatureManagement__{FeatureFlagName.UseRequestWholesaleServicesProcessOrchestration.ToString()}", useRequestWholesaleServicesOrchestration.ToString().ToLower() },
+            { $"FeatureManagement__{FeatureFlagName.UseRequestAggregatedMeasureDataProcessOrchestration.ToString()}", useRequestAggregatedMeasureDataOrchestration.ToString().ToLower() },
         });
     }
